Make HalfCircleDegreeCoordinate non-modular
Fundamentally, the modular behavior of a full circle does
not extend to a half-circle. Most notably, a half-circle
has two distinct poles, and halving the modulus makes these
poles equal to each other. Therefore, our implementation
had a serious bug in that the two points the furthest away
from each other were treated as equal. While it would be
possible to refactor to modulus + 1 and make them non-equal,
there is no benefit to this. Thus, dropping the modulus
entirely is probably a better choice here, as it simplifies
conversions by the formatters that will be implemented. It
therefore is necessary to validate the degree ranges for
the new form of half-circle coordinates.For the full-circle
coordinates, modular behavior is still desirable, and thus
no changes are needed at this time.