Refactor certain comparisons to use 'assertEquals'
Initially, the code was developed using TDD on 'equals' methods. As a
result, many of the assert statements in the tests get a boolean result
from 'equals' and then check whether this is true or false. Though this
is fine for testing the behavior of a boolean method - and arguably
desirable, as a failed test will list expected-actual in boolean terms -
 this is not good practice for any other comparisons of objects. This
 commit refactors all such instances that are not direct testing of an
 'equals' method to use Assertions.assertEquals or
 Assertions.assertNotEquals instead.