Add CustomAssertions::assertBounded
A number of tests for calculating the position of objects and the Sun
will require checking that the predictions are reasonably close to
reality. As this will be handled by asserting the error is bounded, and
as such assertions will occur in multiple classes, it is helpful to
create a static helper method in an appropriately-named class - with
custom messaging displaying the actual deviation - in preparation for
these unit tests.