Fix bug for small negative degree coordinates
By construction, degree coordinates take negatives on the degree
component; however, since -0=0, this makes it impossible to make a
coordinate of a form like -0°1'2" because this will parse to 0°1'2".
This was overlooked because such coordinates amount to less than 1% of
all such coordinates; however, for the very same reason, there is no
need to modify the construction of degree coordinates via the three
components so long as a negation method exists for those rare cases.
However, the parsing of both the CSV files and the command line
necessitates that these be distinct coordinates, and this commit
achieves that by stripping the negative before parsing to integer parts.
 The fix has also been applied to the HourCoordinate converter to avoid
 unexpected behavior as a result of mischievous end-users - although
 it's probably safe to assume that nobody in their right mind would use
 negative hour coordinates when every source is published with purely
 positive hour coordinates, it is best not to count on assumptions.