Revert AskAstraios to pre-'--example'
The last few commits introduced a major problem into the program -
namely, it no longer automatically answers invalid inputs with
appropriate help messages. This is a result of calling parseResult and
not execute - something that fundamentally must be done to check for
'--example' prior to execution. However, even wrapping this call to
defer back to picocli's handlers does not fix this problem, as that will
 cause the '--example' token to be ignored in many of the circumstances
 where that option is most useful. Furthermore, the code required to
 make this work is quite ugly and is not clean conceptually. As a
 result, it is probably best if we discard the '--example' option and
 just direct the users to the 'example' subcommand.